% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{autocorrelation}
\alias{autocorrelation}
\title{Calculate the first order auto-correlation}
\usage{
autocorrelation(x)
}
\arguments{
\item{x}{numeric vector of measurements for one subject}
}
\description{
This function calculate the first order auto-correlation, i.e., the correlation between consecutive observations, for the EMA data in one subject.
A higher value indicates that measured variable (for example, emotion) is sustained over time and shows less homeostatic recovery.
}
\examples{
x <- 1:10
autocorrelation(x)

}
\references{
\insertRef{buu2020association}{EMAanalysis}
}
