% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{get_lambda_o}
\alias{get_lambda_o}
\title{Calculate \eqn{\lambda_o}}
\usage{
get_lambda_o(x)
}
\arguments{
\item{x}{numeric vector of measurements for one subject}
}
\description{
This function calculate \eqn{\lambda_o}, an instability measure of the EMA data in one subject.The data vector is
first standardized. Then we calculate \eqn{\lambda_o} as
\deqn{\lambda_o = n_o/t_o}
where \eqn{n_o} is the number of transitions from outside to inside the standard range, and \eqn{t_o}
is the number of assessments with a score outside the standard range. \eqn{\lambda_o} indicates the conditional probability of transitioning to inside the standard range, given
being outside.
}
\examples{
x <- 1:10
get_lambda_o(x)

}
\references{
\insertRef{buu2020association}{EMAanalysis}
}
