% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{get_lambda_s}
\alias{get_lambda_s}
\title{Calculate \eqn{\lambda_s}}
\usage{
get_lambda_s(x)
}
\arguments{
\item{x}{numeric vector of measurements for one subject}
}
\description{
This function calculate \eqn{\lambda_s}, an instability measure of the EMA data in one subject.The data vector is
first standardized. Then we calculate \eqn{\lambda_s} as
\deqn{\lambda_s = n_s/t_s}
where \eqn{n_s} is the number of transitions from inside to outside the standard range, and \eqn{t_s}
is the number of assessments with a score inside the standard range. \eqn{\lambda_s} measures the conditional probability of transitioning from inside to outside.
}
\examples{
x <- 1:10
get_lambda_s(x)

}
\references{
\insertRef{buu2020association}{EMAanalysis}
}
