% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{summarize_EMA_data}
\alias{summarize_EMA_data}
\title{Get the data summarize for the EMA variable}
\usage{
summarize_EMA_data(ID, x)
}
\arguments{
\item{ID}{a vector contain all the IDs for each measurement in x.}

\item{x}{a vector contain all the measurements for all the subjects.}
}
\description{
This function returns the summary of the EMA variable, which include the mean, variance, autocorrelation,
\eqn{\lambda_o} and \eqn{\lambda_s} for each subject. The function will return a n*6 matrix, where n is the
number of subjects. In each row, we record the subject's ID, and his/her five summary variables listed above.
}
\examples{
ID = rep(seq(1:10), each = 8)
x = rnorm(80)
summarize_EMA_data(ID, x)

}
\references{
\insertRef{buu2020association}{EMAanalysis}
}
